test: guard against float PMS money fields

+++ mhm/src-tauri/src/commands/billing.rs
@@ -34,7 +34,7 @@ pub async fn add_folio_line(
         &booking_id,
         &category,
         &description,
-        amount as f64,
+        amount,
         user_id.as_deref(),
     )
     .await?;
